requirements.txt
python-dotenv
streamlit
pathlib
google-generativeai
youtube_transcript_api